Rename files and functions
