--- src/compile/index.ts
@@ -1,1 +1,0 @@
-export {compile} from './compile';
