--- src/handle-trailing-slash/index.ts
@@ -1,1 +1,0 @@
-export {handleTrailingSlash} from './handle-trailing-slash';
