+++ src/handleTrailingSlash/handleTrailingSlash.ts
