+++ src/queryString/index.ts
@@ -1,0 +1,1 @@
+export {queryString} from './queryString';
