+++ src/queryString/queryString.test.ts
@@ -1,8 +1,8 @@
-import {querystring} from './querystring';
+import {queryString} from './queryString';
 
 describe('Function `querystring`', () => {
     test('Querystring with empty `query` object passed should return null', () => {
-        const received = querystring({});
+        const received = queryString({});
         const expected: null = null;
 
         expect(received).toBe(expected);
@@ -10,7 +10,7 @@ describe('Function `querystring`', () => {
 
     test('Querystring with non-empty `query` object (one pair) passed should return valid querystring', () => {
         const query = { a: 'b' };
-        const received = querystring(query);
+        const received = queryString(query);
         const expected = '?a=b';
 
         expect(received).toBe(expected);
@@ -18,7 +18,7 @@ describe('Function `querystring`', () => {
 
     test('Querystring with non-empty `query` object (several pairs) passed should return valid querystring', () => {
         const query = { a: 'b', c: 'd' };
-        const received = querystring(query);
+        const received = queryString(query);
         const expected = '?a=b&c=d';
 
         expect(received).toBe(expected);
