+++ src/queryString/queryString.ts
@@ -3,7 +3,7 @@
  * @param query Query parameters object
  * @returns Query string
  */
-export const querystring = (query: Record<string, string>): string | null => {
+export const queryString = (query: Record<string, string>): string | null => {
     const entries = Object.entries(query);
 
     return entries.length && '?' + entries.map(pair => pair.join('=')).join('&') || null;
